Add rootReducer, Move initialState to book reducer
