+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import bookReducer from './book';
+
+export default combineReducers({
+  booksObject: bookReducer,
+});
